Refactor code in multiple files to remove cyclic dependencies
